+++ test/protractor/spec/lforms_formdata.spec.js
@@ -204,7 +204,7 @@ describe('get form data', function() {
   it('should not get any data with empty values when the noEmptyValue parameter is used', function() {
     tp.openFullFeaturedForm();
 
-    // only three field has data
+    // only three fields have data
     browser.driver.executeAsyncScript(function () {
       var callback = arguments[arguments.length - 1];
       var fData = LForms.Util.getFormData(null, true, true);